Add files via upload
These are the files associated with my ongoing machine-learning project. This is not complete, however the actual training model does work, I am just having trouble with pipelining my data into the model. Some TensorFlow methods seem to have been deprecated, so I am currently trying to find a way around this. The associated files are used in data generation and configuration of the data to tensors. The actual data files are not included as they were too large